#168 ios navigation with events listener
